Reorganise files into appropriate folder names
